move post logic and create parser folders
