+++ post_service/controllers/post_controller.py
@@ -1,8 +1,7 @@
 from datetime import datetime, timedelta
 
 import requests
-from flask import request
-from flask_restplus import Resource, fields, reqparse, marshal
+from flask_restplus import Resource, fields, marshal
 from sqlalchemy import desc
 
 from post_service.api.restplus import api
@@ -10,6 +9,7 @@ from post_service.models import db
 from post_service.models.category import Category
 from post_service.models.post import Post
 from post_service.models.postvote import Postvote
+from post_service.parsers.post_parsers import *
 
 ns = api.namespace('posts', description='Operations related to posts', path="/<string:category>")
 
@@ -33,33 +33,8 @@ post_dto = api.model('post', {
     'author': fields.Nested(user_dto),
     'new_flag': fields.Boolean(required=True, description='new flag for the post'),
     'edited_flag': fields.Boolean(required=True, description='new flag for the post'),
-    'user_requested_vote': fields.Integer(required=False, description='status of user vote')
+    'vote_type': fields.Integer(required=False, description='status of user vote')
 })
-
-post_get_parser = reqparse.RequestParser()
-post_get_parser.add_argument('requested_user_uuid', required=False, type=str, help='user that requested get')
-
-post_add_parser = reqparse.RequestParser()
-post_add_parser.add_argument('title', required=True, type=str, help='title of post', location='json')
-post_add_parser.add_argument('body', required=True, type=str, help='body of post', location='json')
-post_add_parser.add_argument('image_link', type=str, help='link of attached image', location='json')
-post_add_parser.add_argument('author_uuid', type=str, required=True, help='author uuid', location='json')
-
-post_edit_parser = reqparse.RequestParser()
-post_edit_parser.add_argument('new_title', nullable=True, type=str, help='new title of post', location='json')
-post_edit_parser.add_argument('new_body', nullable=True, type=str, help='new body of post', location='json')
-post_edit_parser.add_argument('new_image_link', nullable=True, type=str, help='new image link', location='json')
-
-post_vote_add_parser = reqparse.RequestParser()
-post_vote_add_parser.add_argument('user_uuid', required=True, type=str, help='uuid of user', location='json')
-post_vote_add_parser.add_argument('vote_type', required=True, type=int, help='vote type', location='json')
-
-post_vote_edit_parser = reqparse.RequestParser()
-post_vote_edit_parser.add_argument('user_uuid', required=True, type=str, help='uuid of user', location='json')
-post_vote_edit_parser.add_argument('new_vote_type', required=True, type=int, help='vote type', location='json')
-
-post_vote_delete_parser = reqparse.RequestParser()
-post_vote_delete_parser.add_argument('user_uuid', required=True, type=str, help='uuid of user', location='json')
 
 
 def get_author(author_uuid):
@@ -74,69 +49,46 @@ def nest_author_info(post):
     return post
 
 
-def get_all_posts():
-    # Calculates 3 days prior to current time
-    three_days_ago = datetime.utcnow() - timedelta(days=3)
-    # Obtains posts from recent 3 days (new posts)
-    new_ordered_posts = Post.query \
-        .filter(Post.pub_date > three_days_ago) \
-        .order_by(desc(Post.upvotes - Post.downvotes)).all()
-    # Obtains posts from prior to recent 3 days (old posts)
-    not_new_ordered_posts = Post.query \
-        .filter(Post.pub_date <= three_days_ago) \
-        .order_by(desc(Post.upvotes - Post.downvotes)).all()
+def get_posts_by_category(category, user_uuid):
+    post_query = Post.query
 
-    # Appends the old posts to the new posts
-    result_posts = new_ordered_posts
-    for post in not_new_ordered_posts:
-        if post is not None:
-            # Inverts new post flag to false as post was queried in old posts
-            post.invert_new_flag()
-            result_posts.append(post)
+    if category != 'all':
+        queried_category = Category.query.filter_by(name=category).first()
+        post_query = post_query.filter_by(category_uuid=queried_category.category_uuid)
 
-    for post in result_posts:
-        nest_author_info(post)
-
-    return result_posts
-
-
-def get_posts_by_category(category):
-    if category == "all":
-        return get_all_posts()
-
-    queried_category = Category.query.filter_by(name=category).first()
+    filteredPostVote = Postvote.query.filter_by(user_uuid=user_uuid).subquery()
     # Calculates 3 days prior to current time
     three_days_ago = datetime.utcnow() - timedelta(days=3)
     # Obtains posts from recent 3 days with category filter (new posts)
-    new_ordered_posts = Post.query \
-        .filter_by(category_uuid=queried_category.category_uuid) \
-        .filter(Post.pub_date > three_days_ago) \
-        .order_by(desc(Post.upvotes - Post.downvotes)).all()
+    new_ordered_posts = post_query.filter(Post.pub_date > three_days_ago) \
+        .order_by(desc(Post.upvotes - Post.downvotes), desc(Post.pub_date))
+
     # Obtains posts from prior to recent 3 days with category filter (old posts)
-    not_new_ordered_posts = Post.query \
-        .filter_by(category_uuid=queried_category.category_uuid) \
-        .filter(Post.pub_date <= three_days_ago) \
-        .order_by(desc(Post.upvotes - Post.downvotes)).all()
+    not_new_ordered_posts = post_query.filter(Post.pub_date <= three_days_ago) \
+        .order_by(desc(Post.upvotes - Post.downvotes), desc(Post.pub_date))
 
     # Appends the old posts to the new posts
-    result_posts = new_ordered_posts
-    for post in not_new_ordered_posts:
+    result_posts = new_ordered_posts.all()
+    for post in not_new_ordered_posts.all():
+        result_posts.append(post)
         if post is not None:
             post.invert_new_flag()
-            result_posts.append(post)
 
     for post in result_posts:
         nest_author_info(post)
+        get_post_vote(post, user_uuid)
 
     return result_posts
+
 
 def get_post_vote(post, user_uuid):
     vote = Postvote.query.filter_by(post_uuid=post.post_uuid) \
         .filter_by(user_uuid=user_uuid).first()
     if vote:
-        post.user_requested_vote = vote.vote_type
+        post.vote_type = vote.vote_type
     else:
-        post.user_requested_vote = None
+        post.vote_type = None
+
 
 @ns.route('/posts')
 class PostCollection(Resource):
@@ -149,16 +101,10 @@ class PostCollection(Resource):
         """
         args = post_get_parser.parse_args()
         try:
-            if category is None:
-                return {"message": "category not found."}, 201
+            if not Category.query.filter_by(name=category).first():
+                return {"message": "category not found."}, 404
 
-            posts = get_posts_by_category(category)
-            for post in posts:
-                if args['requested_user_uuid']:
-                    get_post_vote(post, args['requested_user_uuid'])
-                else:
-                    post.user_requested_vote = None
-
+            posts = get_posts_by_category(category, args['user_uuid'])
             return marshal(posts, post_dto, envelope='posts'), 200
         except Exception as e:
             return {"message": str(e)}, 500
@@ -198,8 +144,8 @@ class PostItem(Resource):
 
         result_post = Post.query.filter_by(post_uuid=post_uuid).first()
         nest_author_info(result_post)
-        if args['requested_user_uuid']:
-            get_post_vote(result_post, args['requested_user_uuid'])
+        if args['user_uuid']:
+            get_post_vote(result_post, args['user_uuid'])
         else:
             result_post.user_requested_vote = None
 
@@ -226,13 +172,13 @@ class PostItem(Resource):
                 post_to_be_edited.edited_date = datetime.utcnow()
                 post_to_be_edited.edited_flag = True
             else:
-                return {'message': 'post specified not found in database'}, 201
+                return {'message': 'post specified not found in database'}, 404
 
             db.session.commit()
         except Exception as e:
             return {"message": str(e)}, 500
 
-        return {'message': 'post has been edited successfully.'}, 201
+        return {'message': 'post has been edited successfully.'}, 200
 
     def delete(self, post_uuid, category):
         """
@@ -269,8 +215,8 @@ class PostSearch(Resource):
                 result_posts.append(post)
 
         for post in result_posts:
-            if args['requested_user_uuid']:
-                get_post_vote(post, args['requested_user_uuid'])
+            if args['user_uuid']:
+                get_post_vote(post, args['user_uuid'])
             else:
                 post.user_requested_vote = None
 
@@ -288,16 +234,17 @@ class UserPosts(Resource):
         args = post_get_parser.parse_args()
 
         result_posts = Post.query.filter_by(author_uuid=user_uuid) \
-                            .order_by(desc(Post.pub_date)).all()
+            .order_by(desc(Post.pub_date)).all()
 
         for post in result_posts:
             nest_author_info(post)
-            if args['requested_user_uuid']:
-                get_post_vote(post, args['requested_user_uuid'])
+            if args['user_uuid']:
+                get_post_vote(post, args['user_uuid'])
             else:
                 post.user_requested_vote = None
 
         return result_posts
+
 
 @ns.route('/<string:post_uuid>/vote')
 class PostVote(Resource):
@@ -334,7 +281,7 @@ class PostVote(Resource):
             new_vote_type = args['new_vote_type']
             post_to_be_edited = Post.query.filter_by(post_uuid=post_uuid).first()
             vote_to_be_edited = Postvote.query.filter_by(post_uuid=post_uuid) \
-                                                .filter_by(user_uuid=user_uuid).first()
+                .filter_by(user_uuid=user_uuid).first()
 
             if vote_to_be_edited and post_to_be_edited:
                 if vote_to_be_edited.vote_type == new_vote_type:
@@ -350,7 +297,6 @@ class PostVote(Resource):
 
         except Exception as e:
             return {"message": str(e)}, 500
-
 
     @ns.expect(post_vote_delete_parser)
     def delete(self, post_uuid, category):
